+++ app/assets/javascripts/discourse-common/addon/lib/icon-library.js
@@ -2,7 +2,7 @@ import I18n from "I18n";
 import attributeHook from "discourse-common/lib/attribute-hook";
 import { h } from "virtual-dom";
 import { isDevelopment } from "discourse-common/config/environment";
-import { escapeExpression } from "discourse/lib/utilities";
+import escape from "discourse-common/utils/escape";
 
 const SVG_NAMESPACE = "http://www.w3.org/2000/svg";
 let _renderers = [];
@@ -141,25 +141,23 @@ registerIconRenderer({
   name: "font-awesome",
 
   string(icon, params) {
-    const id = escapeExpression(handleIconId(icon));
-    let html = `<svg class='${escapeExpression(
-      iconClasses(icon, params)
-    )} svg-string'`;
+    const id = escape(handleIconId(icon));
+    let html = `<svg class='${escape(iconClasses(icon, params))} svg-string'`;
 
     if (params.label) {
       html += " aria-hidden='true'";
     }
     html += ` xmlns="${SVG_NAMESPACE}"><use xlink:href="#${id}" /></svg>`;
     if (params.label) {
-      html += `<span class='sr-only'>${escapeExpression(params.label)}</span>`;
+      html += `<span class='sr-only'>${escape(params.label)}</span>`;
     }
     if (params.title) {
-      html = `<span class="svg-icon-title" title='${escapeExpression(
+      html = `<span class="svg-icon-title" title='${escape(
         I18n.t(params.title)
       )}'>${html}</span>`;
     }
     if (params.translatedtitle) {
-      html = `<span class="svg-icon-title" title='${escapeExpression(
+      html = `<span class="svg-icon-title" title='${escape(
         params.translatedtitle
       )}'>${html}</span>`;
     }
@@ -180,7 +178,7 @@ registerIconRenderer({
         h("use", {
           "xlink:href": attributeHook(
             "http://www.w3.org/1999/xlink",
-            `#${escapeExpression(id)}`
+            `#${escape(id)}`
           ),
           namespace: SVG_NAMESPACE,
         }),
